refactored ASN.1 date/time types by moving all the functionality into base class.
