+++ Asn1Parser/Universal/Asn1GeneralizedTime.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Globalization;
-using SysadminsLV.Asn1Parser.Utils;
 
 namespace SysadminsLV.Asn1Parser.Universal;
 
@@ -14,7 +12,7 @@ public sealed class Asn1GeneralizedTime : Asn1DateTime {
     /// Initializes a new instance of the <strong>Asn1GeneralizedTime</strong> class from a date time object
     /// to encode and value that indicates whether to include millisecond information.
     /// </summary>
-    /// <param name="time">A <see cref="DateTime"/> object.</param>
+    /// <param name="time">A <see cref="DateTime"/> object in local time zone.</param>
     /// <param name="preciseTime">
     /// <strong>True</strong> if encoded value should contain millisecond information, otherwise <strong>False</strong>.
     /// </param>
@@ -23,14 +21,15 @@ public sealed class Asn1GeneralizedTime : Asn1DateTime {
     /// Initializes a new instance of the <strong>Asn1GeneralizedTime</strong> class from a date time object
     /// to encode, time zone information and value that indicates whether to include millisecond information.
     /// </summary>
-    /// <param name="time">A <see cref="DateTime"/> object.</param>
+    /// <param name="time">
+    ///     A <see cref="DateTime"/> object in destination time zone if zone information is provided.
+    ///     Local time zone is assumed if zone information is not provided (null).
+    /// </param>
     /// <param name="zone">A <see cref="TimeZoneInfo"/> object that represents time zone information.</param>
     /// <param name="preciseTime">
     /// <strong>True</strong> if encoded value should contain millisecond information, otherwise <strong>False</strong>.
     /// </param>
-    public Asn1GeneralizedTime(DateTime time, TimeZoneInfo zone = null, Boolean preciseTime = false) : base(TYPE) {
-        m_encode(time, zone, preciseTime);
-    }
+    public Asn1GeneralizedTime(DateTime time, TimeZoneInfo? zone = null, Boolean preciseTime = false) : base(TYPE, time, zone, preciseTime) { }
     /// <summary>
     /// Initializes a new instance of the <strong>Asn1GeneralizedTime</strong> class from an existing
     /// <see cref="Asn1Reader"/> object.
@@ -39,9 +38,7 @@ public sealed class Asn1GeneralizedTime : Asn1DateTime {
     /// <exception cref="Asn1InvalidTagException">
     /// The current state of <strong>ASN1</strong> object is not Generalized Time.
     /// </exception>
-    public Asn1GeneralizedTime(Asn1Reader asn) : base(asn, TYPE) {
-        m_decode(asn.GetTagRawData());
-    }
+    public Asn1GeneralizedTime(Asn1Reader asn) : base(asn, TYPE) { }
     /// <summary>
     /// Initializes a new instance of the <strong>Asn1GeneralizedTime</strong> class from a byte array that
     /// represents encoded UTC time.
@@ -51,21 +48,4 @@ public sealed class Asn1GeneralizedTime : Asn1DateTime {
     /// The current state of <strong>ASN1</strong> object is not Generalized Time.
     /// </exception>
     public Asn1GeneralizedTime(Byte[] rawData) : this(new Asn1Reader(rawData)) { }
-
-    void m_encode(DateTime time, TimeZoneInfo zone, Boolean preciseTime) {
-        Value = time;
-        ZoneInfo = zone;
-        Initialize(new Asn1Reader(Asn1Utils.Encode(DateTimeUtils.Encode(time, zone, false, preciseTime), TYPE)));
-    }
-    void m_decode(Byte[] rawData) {
-        var asn = new Asn1Reader(rawData);
-        Initialize(asn);
-        Value = DateTimeUtils.Decode(asn, out TimeZoneInfo zoneInfo);
-        ZoneInfo = zoneInfo;
-    }
-
-    /// <inheritdoc/>
-    public override String GetDisplayValue() {
-        return Value.ToString(CultureInfo.InvariantCulture);
-    }
 }
